+++ test/ui-testing/license-terms.js
@@ -159,6 +159,7 @@ module.exports.test = (uiTestCtx) => {
             .wait(`#plugin-find-license-modal [role="row"][data-label*="${l.name}"]`)
             .click(`#plugin-find-license-modal [role="row"][data-label*="${l.name}"]`)
             .wait(`#linkedLicenses-remoteId-${i}-license-card`)
+            .waitUntilNetworkIdle(2000)
             .insert(`#linkedLicenses-note-${i}`, l.note)
             .then(done)
             .catch(done);