+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change history for ui-agreements
 
+## 3.2 IN PROGRESS
+* Show/hide interface credentials. Fixes ERM-263
+* Fixed package content lists being truncated at 1000 entries. ERM-409
+* Fixed bug where not all the agreement lines were being fetched.
+* Fixed bug where resetting query term wouldn't execute a search. ERM-426
+* Added ability to filter e-resources by source KB. ERM-299
+* Added agreement filtering by internal contact. ERM-421
+
 ## 3.1.1 2019-08-21
 * Use locally-defined saveAndClose translation key.